Add hot reloading
