+++ package.json
@@ -10,9 +10,12 @@
   "devDependencies": {
     "babel-core": "^5.2.17",
     "babel-loader": "^5.0.0",
-    "webpack": "^1.8.11"
+    "react-hot-loader": "^1.2.7",
+    "webpack": "^1.8.11",
+    "webpack-dev-server": "^1.10.1"
   },
   "scripts": {
+    "start": "node server.js",
     "build": "webpack -w -d",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
